+++ config/routes.php
@@ -54,23 +54,34 @@ Router::scope('/', ['plugin' => 'MeYoutube'], function ($routes) {
 	$routes->connect('/videos', ['controller' => 'Videos', 'action' => 'index'], ['_name' => 'videos']);
 	$routes->connect('/videos/rss', ['controller' => 'Videos', 'action' => 'rss', '_ext' => 'rss'], ['_name' => 'videos_rss']);
 	$routes->connect('/videos/search', ['controller' => 'Videos', 'action' => 'search'], ['_name' => 'videos_search']);
-	$routes->connect('/videos/:year/:month/:day', ['controller' => 'Videos', 'action' => 'index_by_date'], [
-		'_name'	=> 'videos_by_date',
+	$routes->connect('/videos/:year/:month/:day', ['controller' => 'Videos', 'action' => 'index_by_day'], [
+		'_name'	=> 'videos_by_day',
 		'year'	=> '[12][0-9]{3}',
 		'month'	=> '0[1-9]|1[012]',
 		'day'	=> '0[1-9]|[12][0-9]|3[01]',
 		'pass'	=> ['year', 'month', 'day']
 	]);
+    $routes->connect('/videos/:year/:month', ['controller' => 'Videos', 'action' => 'index_by_month'], [
+		'_name'	=> 'videos_by_month',
+		'year'	=> '[12][0-9]{3}',
+		'month'	=> '0[1-9]|1[012]',
+		'pass'	=> ['year', 'month']
+	]);
+    $routes->connect('/videos/:year/', ['controller' => 'Videos', 'action' => 'index_by_year'], [
+		'_name'	=> 'videos_by_year',
+		'year'	=> '[12][0-9]{3}',
+		'pass'	=> ['year']
+	]);
 	$routes->connect('/videos/today', [
 		'controller'	=> 'Videos', 
-		'action'		=> 'index_by_date',
+		'action'		=> 'index_by_day',
 		'year'			=> date('Y'),
 		'month'			=> date('m'),
 		'day'			=> date('d'),
 	], ['_name' => 'videos_today', 'pass' => ['year', 'month', 'day']]);
 	$routes->connect('/videos/yesterday', [
 		'controller'	=> 'Videos', 
-		'action'		=> 'index_by_date',
+		'action'		=> 'index_by_day',
 		'year'			=> (new Time('1 days ago'))->i18nFormat('YYYY'),
 		'month'			=> (new Time('1 days ago'))->i18nFormat('MM'),
 		'day'			=> (new Time('1 days ago'))->i18nFormat('dd'),