+++ src/Controller/VideosCategoriesController.php
@@ -46,7 +46,7 @@ class VideosCategoriesController extends AppController {
 	
 	/**
 	 * Lists videos for a category.
-     * It uses the `index` template.
+     * It uses the `Videos/index` template.
 	 * @param string $category Category slug
 	 */
 	public function view($category = NULL) {
@@ -68,7 +68,7 @@ class VideosCategoriesController extends AppController {
 					'Categories' => ['fields' => ['title', 'slug']],
 					'Users' => ['fields' => ['first_name', 'last_name']],
 				])
-				->select(['id', 'youtube_id', 'title', 'subtitle', 'description', 'created'])
+				->select(['id', 'youtube_id', 'title', 'subtitle', 'text', 'created'])
 				->where(['Categories.slug' => $category, 'is_spot' => FALSE])
 				->order([sprintf('%s.created', $this->VideosCategories->Videos->alias()) => 'DESC']);
 					
@@ -88,6 +88,6 @@ class VideosCategoriesController extends AppController {
         
 		$this->set(compact('videos'));
 		
-		$this->render('index');
+		$this->render('Videos/index');
 	}
 }